YT-18571: Extract various invoker wrappers from action_queue.h/cpp commit_hash:6456c8216e462ebffd1c52d01ab3846db4162896
