+++ yt/yt/core/ya.make
@@ -56,6 +56,7 @@ SRCS(
     concurrency/async_semaphore.cpp
     concurrency/async_stream_helpers.cpp
     concurrency/async_stream_pipe.cpp
+    concurrency/bounded_concurrency_invoker.cpp
     concurrency/config.cpp
     GLOBAL concurrency/configure_fiber_manager.cpp
     concurrency/coroutine.cpp
@@ -80,12 +81,15 @@ SRCS(
     concurrency/periodic_executor.cpp
     concurrency/periodic_yielder.cpp
     concurrency/pollable_detail.cpp
+    concurrency/prioritized_invoker.cpp
     concurrency/profiling_helpers.cpp
     concurrency/propagating_storage.cpp
     concurrency/quantized_executor.cpp
     concurrency/scheduler_thread.cpp
+    concurrency/serialized_invoker.cpp
     concurrency/single_queue_scheduler_thread.cpp
     concurrency/suspendable_action_queue.cpp
+    concurrency/suspendable_invoker.cpp
     concurrency/system_invokers.cpp
     concurrency/thread_affinity.cpp
     concurrency/thread_pool_detail.cpp
@@ -93,6 +97,7 @@ SRCS(
     concurrency/thread_pool.cpp
     concurrency/throughput_throttler.cpp
     concurrency/two_level_fair_share_thread_pool.cpp
+    concurrency/watchdog_invoker.cpp
     concurrency/retrying_periodic_executor.cpp
     concurrency/scheduled_executor.cpp
